[PH2][Settings] Refactor and create new file
