+++ Package.swift
@@ -420,6 +420,8 @@ let package = Package(
         "src/core/ext/transport/chttp2/transport/http2_settings.h",
         "src/core/ext/transport/chttp2/transport/http2_settings_manager.cc",
         "src/core/ext/transport/chttp2/transport/http2_settings_manager.h",
+        "src/core/ext/transport/chttp2/transport/http2_settings_promises.cc",
+        "src/core/ext/transport/chttp2/transport/http2_settings_promises.h",
         "src/core/ext/transport/chttp2/transport/http2_stats_collector.cc",
         "src/core/ext/transport/chttp2/transport/http2_stats_collector.h",
         "src/core/ext/transport/chttp2/transport/http2_status.h",
