+++ src/python/grpcio/grpc_core_dependencies.py
@@ -163,6 +163,7 @@ CORE_SOURCE_FILES = [
     'src/core/ext/transport/chttp2/transport/http2_client_transport.cc',
     'src/core/ext/transport/chttp2/transport/http2_settings.cc',
     'src/core/ext/transport/chttp2/transport/http2_settings_manager.cc',
+    'src/core/ext/transport/chttp2/transport/http2_settings_promises.cc',
     'src/core/ext/transport/chttp2/transport/http2_stats_collector.cc',
     'src/core/ext/transport/chttp2/transport/http2_transport.cc',
     'src/core/ext/transport/chttp2/transport/huffsyms.cc',
