+++ tools/doxygen/Doxyfile.core.internal
@@ -1222,6 +1222,8 @@ src/core/ext/transport/chttp2/transport/http2_settings.cc \
 src/core/ext/transport/chttp2/transport/http2_settings.h \
 src/core/ext/transport/chttp2/transport/http2_settings_manager.cc \
 src/core/ext/transport/chttp2/transport/http2_settings_manager.h \
+src/core/ext/transport/chttp2/transport/http2_settings_promises.cc \
+src/core/ext/transport/chttp2/transport/http2_settings_promises.h \
 src/core/ext/transport/chttp2/transport/http2_stats_collector.cc \
 src/core/ext/transport/chttp2/transport/http2_stats_collector.h \
 src/core/ext/transport/chttp2/transport/http2_status.h \
